Extract duplicate code into SearchEndpointE2ESpecHelper
